Introduced ListProductQueryHelper
